Move video player into more robust general component
